Add shared note pool for active presets with increased arp/seq limits
Replace fixed-size inline note arrays with a shared RAM pool (2048 notes,
6144 bytes) to support larger preset limits without wasting memory:
- Arp: max 256 notes / 128 steps (was 64 notes / 127 steps)
- Seq: max 512 notes / 256 steps (was 128 notes / 127 steps)

Key changes:
- New active_preset_t type with uint16_t note_count/pattern_length and
  pointer-based notes from shared pool (replaces direct use of EEPROM types)
- Pool management: bump allocator with compaction, 13 slots (1 arp +
  8 seq + 4 arp QB)
- Dual timing format: 7-bit (EEPROM/factory/HID) and 8-bit (active RAM),
  with NOTE_CONVERT_7TO8 macro for load-time conversion
- EEPROM types/format and HID protocol remain unchanged
- Factory presets use NOTE7_PACK (7-bit), active code uses NOTE_PACK_TIMING_VEL (8-bit)
- Seq update optimized: inline note processing eliminates 2KB+ stack arrays
- All note iteration loops widened to uint16_t
- Helper functions (get_ms_per_16th, seq_get_ms_per_16th, etc.) updated
  to accept active_preset_t*

+++ vial-qmk - ryzen/keyboards/orthomidi5x14/orthomidi5x14.c
@@ -17062,7 +17062,7 @@ void render_quick_build_recording(void) {
     oled_write_line(2, "");
 
     // Line 3: Step number (centered)
-    uint8_t step = quick_build_get_current_step();
+    uint16_t step = quick_build_get_current_step();
     snprintf(buf, sizeof(buf), "STEP %d", step);
     oled_write_line_centered(3, buf);
